backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
google-generativeai==0.3.2
playwright==1.40.0
httpx==0.25.2
python-dotenv==1.0.0
praw==7.7.1
beautifulsoup4==4.12.2
lxml==4.9.3
python-multipart==0.0.6
aiofiles==23.2.1
